Add CLI parser using StructOpt
Kind of would like to add sanity tests for this that just check if the
help message is printed etc. but we'd have to do environment variable
manipulation etc. so that the files don't actually get created in your
home directory. It's not too hard and I've done it before but I kind of
want to implement some other stuff first.